TEMA-3/requirements.txt
Flask==2.3.3
pyodbc==5.3.0
gunicorn==20.1.0
opencensus-ext-azure==1.1.13
opencensus-ext-flask==0.8.1
opencensus-ext-logging==0.1.1